+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 * Do not take user to Orders tab after edit of POL in POL tab. Refs UIOR-799.
 * Filter label contains extra 's'. Refs UIOR-802.
 * Settings (Order templates) | Apply baseline keyboard shortcuts. Refs UIOR-801.
+* Create keyboard shortcut for "Add POL" action. Refs UIOR-812.
 
 ## [2.4.2](https://github.com/folio-org/ui-orders/tree/v2.4.2) (2021-09-08)
 [Full Changelog](https://github.com/folio-org/ui-orders/compare/v2.4.1...v2.4.2)